aula-05/pyproject.toml
[project]
name = "aula-05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standart]>=0.115.11",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pymongo[srv]>=4.11.2",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
]
